Rename Servo.h to ESP32ArduinoServo.h
This will help the library avoid naming clashes with the
default Arduino servo library.
