+++ examples/02-ServoPotentiometer/02-ServoPotentiometer.ino
@@ -1,4 +1,4 @@
-#include <Servo.h>
+#include <ESP32ArduinoServo.h>
 
 static const int servoPin = 4;
 static const int potentiometerPin = 32;
