+++ src/ESP32ArduinoServo.cpp
@@ -30,7 +30,7 @@
  * 2017 Jul 5: Edited by Jaroslav Páral (jarekparal) - paral@robotikabrno.cz
  */
 
-#include <Servo.h>
+#include <ESP32ArduinoServo.h>
 
 int Servo::channel_next_free = 0;
